Use GNU parallel to speed up circleci build
Looking at the circleCI site, I can see that the bulk of the time for
the build is spent installing all of the different python versions
under pyenv. We can use the GNU parallel command to easily run all of
these installs at the same time. I have verified that the parallel
command is available in the circleci python image. Should help with #4.